broadcast/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
dj-database-url==0.5.0
dj-email-url==1.0.5
Django==4.0.4
django-cache-url==3.4.0
django-embed-video==1.4.3
environs==9.5.0
idna==3.3
marshmallow==3.15.0
packaging==21.3
pyparsing==3.0.8
python-dotenv==0.20.0
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.9
